able to hide logout button when logout is clicked
